requirements.txt

python-telegram-bot==13.17
feedparser==6.0.10
apscheduler==3.10.6
Flask==2.3.2
